♻️ (server,client): Refactor auth related code
