+++ apps/client/components/GetUser/GetUsersViewController.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useQuery, gql } from '@apollo/client';
 import { GetUsersViewModel } from './useGetUsersViewModel.query';
 import { ViewController } from '@readable/core/ViewController';
-import { serverHost } from '@readable/link';
+import { serverHost } from '@readable/common/link';
 
 export const GetUsersViewController: ViewController<GetUsersViewModel> = React.memo(({ viewModel }) => {
   const { users, loading, error } = viewModel;
