+++ apps/server/src/auth/auth.controller.ts
@@ -18,7 +18,7 @@ export class AuthController {
     const token = await this.authService.googleLogin(req);
 
     if (token) {
-      res.redirect(`${process.env.CLIENT_HOST}/token?token=${token}`);
+      res.redirect(`${process.env.CLIENT_HOST}/auth?token=${token}`);
     }
   }
 
@@ -32,7 +32,7 @@ export class AuthController {
     const token = await this.authService.githubLogin(req);
 
     if (token) {
-      res.redirect(`${process.env.CLIENT_HOST}/token?token=${token}`);
+      res.redirect(`${process.env.CLIENT_HOST}/auth?token=${token}`);
     }
   }
 
@@ -46,7 +46,7 @@ export class AuthController {
     const token = await this.authService.facebookLogin(req);
 
     if (token) {
-      res.redirect(`${process.env.CLIENT_HOST}/token?token=${token}`);
+      res.redirect(`${process.env.CLIENT_HOST}/auth?token=${token}`);
     }
   }
